Print the first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int n = 50, i;
-	long long int fib1 = 1, fib2 = 2, next;
+	long int fib1 = 1, fib2 = 2, next;
 
 	printf("%ld, %ld, ", fib1, fib2);
